refactor: 네이밍 변경 -카멜케이스
