+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import HomeTemplate from '@/templates/Hometemplate'
+import HomeTemplate from '@/templates/HomeTemplate'
 import { GetServerSideProps } from 'next'
 import { Board } from '@/service/getBoards/boards.type'
 import { getBoards } from '@/service/getBoards/getBoards'
@@ -7,11 +7,11 @@ interface HomePageProps {
   boards: Board[]
 }
 
-export default function Home({ boards }: HomePageProps) {
-  return <HomeTemplate initialData={boards} />
+export default function Home({boards}: HomePageProps) {
+  return <HomeTemplate initialData={boards}/>
 }
 
-export const getServerSideProps: GetServerSideProps = async ({ res }) => {
+export const getServerSideProps: GetServerSideProps = async ({res}) => {
   res.setHeader(
     'Cache-Control',
     'public, s-maxage=10, stale-while-revalidate=59'
@@ -26,7 +26,7 @@ export const getServerSideProps: GetServerSideProps = async ({ res }) => {
     }
 
     return {
-      props: { boards },
+      props: {boards},
     }
   } catch (error) {
     throw error
