+++ src/templates/HomeTemplate/HomeTemplate.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import * as Styles from './styles'
-import CardList from '@/templates/Hometemplate/CardList'
+import CardList from '@/templates/HomeTemplate/CardList'
 import { Board } from '@/service/getBoards/boards.type'
 
 interface HomeTemplateProps {
   initialData: Board[]
 }
 
-const HomeTemplate = ({ initialData }: HomeTemplateProps) => {
+const HomeTemplate = ({initialData}: HomeTemplateProps) => {
   return (
     <Styles.Wrapper>
-      <CardList boards={initialData} />
+      <CardList boards={initialData}/>
     </Styles.Wrapper>
   )
 }
